Sets data in now persisted as displayed
No more persisting data when navigating up. Instead, a new set is persisted in the db when clicking the "new set" button, a set is updated in the db when the focus is lost on one of its textFields, and a set is deleted from the db when clicking on the "delete set" button.