Replaced BottomSheet by a ModalBottomSheet
Better fits the need of adding a new block.

Note : Had to upgrade the M3 plugin version to 1.2.0-alpha02 to fix IME keyboard not showing when clicking on the text field embedded in the ModalBottomSheet. No using alpha03 because that causes some errors with Android API version not being supported