Deleting a block right after a first one was deleted now dismisses the snackbar and displays a new one
This wasn't the case. Instead, the second snackbar was displayed only after the first one was finished.
This was fixed by executing the code displaying the snackbar in a coroutine scope instead of the main thread. Also, before displaying a snackbar, we make sure to dismiss any snackbar that would already be displayed